it's ok

+++ server/sqlhelp/blog.sql
@@ -369,7 +369,7 @@ CREATE TABLE  IF NOT EXISTS blog_manager (
  m_password varchar(255) NOT NULL COMMENT '密码',
  m_token varchar(255) NOT NULL COMMENT '登录token',
  m_group tinyint(3) NOT NULL COMMENT '管理员组',
- m_last_login_time int(13) NOT NULL COMMENT '上次登录时间',
+ m_last_login_time BIGINT(15) NOT NULL COMMENT '上次登录时间',
  m_login_times int(6) DEFAULT 0 COMMENT '登录次数',
  m_head varchar(100) NOT NULL COMMENT '头像',
  PRIMARY KEY (m_id)
